Cleanup old code that is now in the accessqueue pkg and update tests.
